+++ mono/tests/finalizer-exception.cs
@@ -1,39 +1,24 @@
-
-using System; 
-using System.Collections; 
+using System;
 using System.Threading;
 
-public class foo  { 
-	public static LocalDataStoreSlot dataslot = Thread.AllocateDataSlot();
-	public static int final_count=0;
+public class FinalizerException {
+	~FinalizerException () {
+		throw new Exception ();
+	}
 
-	~foo() { 
-		// Demonstrate that this is still the same thread
-		string ID=(string)Thread.GetData(dataslot);
-		if(ID==null) {
-			Console.WriteLine("Set ID: foo");
-			Thread.SetData(dataslot, "foo");
-		}
+	public static int Main () { 
+		AppDomain.CurrentDomain.UnhandledException += (sender, args) => {
+			Console.WriteLine ("caught");
+			Environment.Exit (0);
+		};
 
-		// Don't run forever
-		if(final_count++>10) {
-			Environment.Exit(0);
-		}
+		new FinalizerException ();
 
-		Console.WriteLine("finalizer thread ID: {0}", (string)Thread.GetData(dataslot));
-		throw new SystemException("wibble");
-	} 
+		GC.Collect ();
+		GC.WaitForPendingFinalizers ();
 
-	public static int Main() { 
-		ArrayList list = new ArrayList (); 
-		Thread.SetData(dataslot, "ID is wibble");
-		Environment.ExitCode = 2;
-		while(true) { 
-			foo instance = new foo(); 
-			list.Add (new WeakReference(instance)); 
-			Thread.Sleep (0);
-		}
-		return 1;
-	} 
-} 
+		Thread.Sleep (Timeout.Infinite); // infinite wait so we don't race against the unhandled exception callback
 
+		return 2;
+	}
+}
